+++ ext/log-processor/src/main/java/core/DiagramModule.java
@@ -1,11 +1,10 @@
 package core;
 
+import core.diagram.service.DiagramService;
+import core.diagram.web.DiagramController;
+import core.diagram.web.DiagramModel;
 import core.framework.http.HTTPMethod;
 import core.framework.module.Module;
-import core.visualization.service.ActionFlowService;
-import core.visualization.web.DiagramAJAXService;
-import core.visualization.web.DiagramAJAXServiceImplV2;
-import core.visualization.web.DiagramController;
 
 /**
  * @author neo
@@ -13,8 +12,10 @@ import core.visualization.web.DiagramController;
 public class DiagramModule extends Module {
     @Override
     protected void initialize() {
-        bind(ActionFlowService.class);
-        api().service(DiagramAJAXService.class, bind(DiagramAJAXServiceImplV2.class));
-        http().route(HTTPMethod.GET, "/vis", bind(DiagramController.class)::home);
+        bind(DiagramService.class);
+
+        site().template("/template/diagram.html", DiagramModel.class);
+        DiagramController controller = bind(DiagramController.class);
+        http().route(HTTPMethod.GET, "/diagram/arch", controller::arch);
     }
 }