+++ ext/log-processor/src/main/java/core/diagram/service/Diagram.java
@@ -1,7 +1,6 @@
-package core.visualization.web;
+package core.diagram.service;
 
 import core.framework.api.json.Property;
-import core.framework.api.validate.NotNull;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -9,21 +8,14 @@ import java.util.List;
 /**
  * @author allison
  */
-public class DiagramResponse {
-    @NotNull
-    @Property(name = "graph")
-    public String graph;
-
-    @NotNull
-    @Property(name = "tooltips")
+public class Diagram {
+    public String dot;
     public List<Tooltip> tooltips = new ArrayList<>();
 
     public static class Tooltip {
-        @NotNull
         @Property(name = "id")
         public String id;
 
-        @NotNull
         @Property(name = "html")
         public String html;
     }