+++ CHANGELOG.md
@@ -1,10 +1,13 @@
 ## Change log
 
 ### 7.6.14 (03/18/2021 - )
+
 * search: update es to 7.12.0
 * action: change error "action log context value too long" to warning
 * mongo: MongoMigration supports overwriting property by env
 * es: ElasticSearchMigration supports overwriting property by env
+* log-processor: generate system arch diagram
+  > call https://log-processor:8443/diagram/arch, use port-forward in kube env
 
 ### 7.6.13 (03/01/2021 - 03/18/2021)